arch: Remove architecture library from CI script
Much of how the architecture library behaves depends specifically on the
configuration of the target platform, and this configuration comes from
the firmware. In future patches, using this information will prevent the
architecture library from being buildable in isolation.

In preparation for this, the architecture library has been removed from
the CI script, and we will instead rely on its successful build within
each platform to gauge success.

Change-Id: I6740d2c3a828e4a1c29a276a07095d809ec62e33
Signed-off-by: Chris Kay <chris.kay@arm.com>