arch: Move interrupt numbers to CMSIS configuration
Because we compile with `-fshort-enums`, the current CMSIS IRQ
enumeration, which only holds the exception numbers, is reduced in width
to a single 8-bit value. This means that any attempt to convert from a
large interrupt number to an IRQn_Type, which CMSIS does in the next
patch, is not safe.

We fix this by moving the interrupt numbers to the IRQn_Type, so that we
can use the CMSIS APIs to manage these interrupts.

Change-Id: Ic2644b5b97fb6dab412d50b8168597c056727295
Signed-off-by: Chris Kay <chris.kay@arm.com>