misc: Replace `fwk_assert(false)` with `fwk_unexpected()`
This commit replaces any calls to `fwk_assert(false)` with a call to
`fwk_unexpected()`. Forced assertions are used for two different
scenarios:

- Unreachable code, which should use `fwk_unreachable()`
- Code that should trap only in debug mode, which should use
  `fwk_unexpected()`

`fwk_unexpected()` is the only one of these options with no side effects
that did not already occur in the original code, so these assertions
have been transformed to expectations. Some of these may, and should, be
transformed later to unreachable code.

Change-Id: I8325259b41a03a096b62a420bd0c7d52b025fcf4
Signed-off-by: Chris Kay <chris.kay@arm.com>