rddaniel: mscp_rom: set vector table offset register
With the use of CMSIS from Commit afd75a9 ("Refactor the architecture
layer"), ROM firmware is expected to set the vector table offset
register to the RAM firmware's vector table before jumping to the RAM
firmware. This patch sets the VTOR register to the RAM firmware's vector
table for rddaniel.

Change-Id: I4673acebe5e7039c0a4694ea3f4e694f342bb821
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>