rddanielxlr: mscp_rom: set vector table offset register
With the use of CMSIS from Commit afd75a9 ("Refactor the architecture
layer"), ROM firmware is expected to set the vector table offset
register to the RAM firmware's vector table before jumping to the RAM
firmware. This patch sets the VTOR register to the RAM firmware's vector
table for rddanielxlr platform.

Change-Id: I29d36131274bb32b315fcd48564ee5ea4196ea1d
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>