fwk: Add compiler attribute macros
Some attributes are not available on all the toolchains we support. We
need to have the ability to conditionally remove macros and hide the
attribute syntax from the user for attributes that are not universally
supported.

This commit adds a macro alternative to all currently used attributes
within the project to date.

Change-Id: Ibf84bc8d0dc941c9db2f946e4ac20dd5bf44bf24
Signed-off-by: Chris Kay <chris.kay@arm.com>