juno/config_psu: Fix PSU driver identifiers
At present, the driver identifiers for all of Juno's PSU elements
incorrectly point to the first Mock PSU driver element, instead of to
their respective elements. This causes issues when trying to interface
with the mock PSUs because they modify a single global state instead of
independent states.
This commit resolves the mismatch in driver identifiers so that they
point to their respective Mock PSU elements.

Change-Id: I6f486ee11467348453eaa4d046c6411d9c458941
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>