framework: set event count from notification count
The default value of EVENT_COUNT is set to 64. On platforms that
requires larger number of notification subscriptions (192 for example),
EVENT_COUNT of 64 will not be sufficient. When platform sets the
max notification count greater than 64, set the EVENT_COUNT to
max notification count to account for additional events due to increased
notifications.

Change-Id: I9370703b47268c08d3f8bfb32b21f07b8b5370f7
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>