fwk/assert: print location of assertion
fwk_assert which is called from several framework and module functions,
enters trap loop silently if condition passed to it fails, without
providing any information about the failure in the debug builds. Such
cases will be hard to debug in case of unavailability of a debugger or
textual trace output. Print the file name and line number to know the
location of assertion before entering infinite loop.

Change-Id: I9ff067178373c8c38ecfd25d9efc322b3d393ed3
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>