mod_scmi_pd: Fix invalid arguments
The SCMI v2 specification defines the bits 1-31 of the flags
parameter for power_state_set to be restricted, hence the SCP
should answer with invalid parameters, whenever these are found
set. This is the similar case for the bits 31, 29-28 of the
power_state parameter. This commit renders the SCP conformant to
the SCMI Test Suite 2.0 Power Domain Test 7, which also caused
the PPU to hang when setting the restricted bits.

Change-Id: I936965af9385ed8b48db6e185218610eac6c1ae1
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>