fwk: Initialize module context structures early
At present, module and element context structures are initialized only
once the module component has been explicitly started by the
architecture layer. This presents problems when trying to use context
data before the module system has come online.

With the introduction of the I/O framework component later on,
components of the framework need to be able to access certain module
data before the module itself has come online to, for instance, validate
an identifier.

The introduction of this change greatly increases the availability of
certain framework interfaces before the module system comes up by
separating the context initialization from the system initialization.

To ease debuggability, a number of framework checks critical to the
continuation of the firmware now trap at the point of failure rather
than returning error codes up the stack (at which point a lot of useful
information is lost).

Change-Id: Idfa28cf94cecbc8e3e72108fbe7a8776c814aa07
Signed-off-by: Chris Kay <chris.kay@arm.com>