host: Remove support for multithreading
The current host architecture supports a custom CMSIS-RTOS2 layer that
interfaces with the host system implementation of pthreads.

Given the host product can now be run in single-threaded mode, the extra
maintenance required to keep the multi-threaded version functioning is
prohibitive, especially with frequent changes to the framework.

This commit switches the `host` product's `fw` firmware to
single-threaded mode, and removes support for building the `host`
architecture in multi-threaded mode.

Change-Id: I59a7db0eb0f2bb70614c1fe4aca3f482662976f3
Signed-off-by: Chris Kay <chris.kay@arm.com>