pl011: Refactor and fix module dependencies
This change refactors the PL011 in preparation for future I/O patches.

Aside from general code style updates, this patch fixes a hidden
dependency on the system power module, removes its dependency on the
clock module, and introduces more safety checks around when the device
is powered on and off.

Change-Id: Ic3e87739cfd4a365372310745f737f685cf57455
Signed-off-by: Chris Kay <chris.kay@arm.com>