power_domain: Selectively disable pre and post transition notifications.
This change allows a platform to configure pre and post state
transition notifications to be disabled for a power domain, in a case
for some reason if it is not desired e.g. optimization where
notifications handling code won't execute unnecessarily if there are
no subscribers for the notifications for that domain. By default
pre and post state transition notifications are enabled for all domains.

Change-Id: I9bbff3f2d2702abf1566ea846e70b22fa9c45ea4
Signed-off-by: Girish Pathak <girish.pathak@arm.com>