scmi_perf: Add performance plugins handler extension
There is a need to interleave some performance-related regular
modules between SCMI and the performance driver (DVFS). The reason
for this is to insert some logic, encapsulated in modules, that can
affect the final performance state for the platform.
To enable a scalable way to do so, a layer between the scmi-perf and
DVFS modules is added, providing basic infrastructure to easily add and
remove modules, called performance plugins.

At the same time, this new extension can work as a decoupling layer
between the domains exposed to SCMI from those that are managed by
the DVFS module. So far there's always been an implicit 1:1
mapping, but SCMI has the ability to abstracts such domains.
This change also adds a domain aggregation mechanism by which it is
possible to handle logical domains, exposed by SCMI, mapped to physical
ones, managed by DVFS.
This is to support a case, for example, where the firmware may want to
expose per-cpu domains (or controls) to OS, while the platform has fewer
frequency (physical) domains.
As this patch adds this new domains aggregation, the exact number of
SCMI performance domains (perf_doms_count) must be provided explicitly
and all platforms' configurations have been updated accordingly.

This new layer is optional and can be enabled by choosing
BUILD_HAS_SCMI_PERF_PLUGIN_HANDLER.

Change-Id: Ic6d79802c26120f2c75b23fd4b37c0c23e0baa18
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>