cmake: Add `format`, `lint` and `check` targets
This commit introduces the `format`, `lint` and `check` build system
targets, which can be used to automatically run quality assurance checks
over the code-base. Individual backends for these targets are added
incrementally further up the tree.

The `check` target is unique in that is intended to encompass all
possible checks, including formatting checks, lints and tests.

Change-Id: I6042a602478fb38e1019552c4bacfee91920774a
Signed-off-by: Chris Kay <chris.kay@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>