cmake: Add firmware configuration handling
This commit adds the fundamental firmware selection logic. The user is
expected to pass an initial-cache file to CMake (e.g. with `-C` on the
command line) when they initialize a build directory, and this
initial-cache file is expected to create a set of cache variables
describing the firmware and setting up any configuration options it
might need.

Change-Id: Icd0b8251b13ebb62ba36da53293f9ecced7dca97
Signed-off-by: Chris Kay <chris.kay@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>