cmake: Enable flat binary generation
This commit adds support for generating flat binary images, controlled
by the `SCP_GENERATE_FLAT_BINARY` option. This option defaults to false,
but firmware images that wish to enforce a value for it may internalize
the variable.

The resulting binary is generated with `objdump` from binutils, or from
`fromelf` if the firmware has been compiled with Arm Compiler 6.

Change-Id: I83b90262e196ca8e16cc9e887afb92bfa23dafdd
Signed-off-by: Chris Kay <chris.kay@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>