fwk: Add fwk_thread_signal API
This patch adds the prototype for fwk_thread_signal API.
It also adds a signal handler to the module configuration.

A signal may be used instead of an event where the target module
does not require any parameters and is not required to return any
data to the caller. It is intended as a lightweight supplement to
the standard event model, eliminating most of the overhead of
the framework event processing.

Change-Id: Iac6f08edeaa167b04941454628f0ed592e0aa6f8
Signed-off-by: Jim Quigley <jim.quigley@arm.com>