product/rddanielxlr: migrate to isys_rom
Migrate from using mscp_rom module to isys_rom module as the boot ROM
for both SCP and MCP controllers. As there is no support yet for MCP
RAM firmware, the config data for the bootloader module is set to NULL.
The isys_rom module ensures that the MCP processor is held in a idle
state.

Change-Id: Icfe6f6a0d01c49f63d432d188e9b5c98a492fd47
Signed-off-by: Aditya Angadi <aditya.angadi@arm.com>