misc: Add initial Jenkinsfile
The Jenkinsfile added by this commit represents the core CI pipeline
used by the internal Jenkins server. The pipeline runs as a sandboxed
declarative pipeline, and is executed according to the internal CI
configuration.

For information on the Jenkinsfile and what it does, see the official
documentation: https://www.jenkins.io/doc/book/pipeline/jenkinsfile/

Change-Id: Ia1cdd749958020885cd62b5235300bf5eff9b189
Signed-off-by: Chris Kay <chris.kay@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>