juno/config_scmi_clock: Fix startup clocks state
The configuration of the SCMI clock module for Juno is set to
hold a stopped state for all the configured clock elements.
Nonetheless, Juno actually starts all the clocks at startup,
causing a contradiction in the configuration. This commit fixes
this issue, by setting Juno's clock states to RUNNING at
startup.

Change-Id: Id315c909b06db40ea3b46bd242d05cafbeacbf33
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>