juno/xrp7724: Log out of margin in place of propagating error
Currently when the voltage measured on the rail after a set voltage
operation is out of margin an error is returned. Upon this error, no
further action is taken. This causes the framework to log an error,
while this is rather a driver expectation.

Manual testings have also shown that the threshold is probably too
tight as in some circumstances the deviation from the target reaches
30-35 mV.

Since no action is taken after the margin comparison, simply log the
fact from the driver itself and return FWK_SUCCESS, as the event has
been processed correctly.

Change-Id: I6db0950a9376bec4b29f1d2d08a31e669f76f9c9
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>