mod_scmi_reset: Fix RESET invalid reset_state
As per section 4.8.1 of the SCMI v2 spec, the only valid reset id for
the Architectural reset state type is 0. All the other IDs available in
the 32-bit space are reserved for future use. This commit brings an
early exit condition that returns SCMI_INVALID_PARAMETERS when it
detects that any of those ids are used.

Change-Id: I81d7aee541e7577e58e242da6bebf14edd00b753
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>