mod_scmi_power: Fix invalid device return value
The SCMI power domain protocol does not support notifications
for all the the domains. Specifically, only DEVICE type ones do.
The code originally correctly picked this up and returned with an error,
with the exception that this error was SCMI_DENIED. As per section 4.1.4
of the SCMI v2 spec, this error is dedicated to permissions-related
problems. As this is not the case, this
commit replaces this error with a more suitable SCMI_NOT_SUPPORTED.

Change-Id: I537391bcce57a7849dae2ed26924ca1ad0b1016b
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>