juno/scmi_perf: Simplify configuration of FastChannels
The current configuration for FastChannels in Juno allows a generic
placement of addresses for each of the channels. This, however,
requires a number of definitions that do not scale easily should we
want to expose an arbitrary set of channels.

To simplify, combine the common definitons under some macros, to
improve readability.

No functionality is affected by this change.

Change-Id: I1ed12f4c148b9b23cbe46dd3256a04824367c77d
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>