juno/mod_hdlcd: Fix invalid rate response
With the premise that a clock rate is 64 bits long, this commit
fixes the behaviour where the Juno HDLCD clock driver takes only
the lower 32 bits of the rate ignoring the remaining upper half.
With the assumption that only 32 bits are therefore required,
this commit includes a check whether the upper 32 bits are zero,
enforcing this requirement and returning FWK_E_PARAM if any
provided rate is bigger than 32 bits.

Change-Id: I0c72432a91d9e4a43c44b36e343a884058ebada9
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>