synquacer: scp_romfw: set vector table offset register
With the use of CMSIS from Commit afd75a9 ("Refactor the architecture
layer"), ROM firmware is expected to set the vector table offset
register to the RAM firmware's vector table before jumping to the RAM
firmware. This patch sets the VTOR register to the RAM firmware's vector
table for synquacer platform.

Change-Id: I720fe8249b0d40a56028af0c26d00ef1aa2156d2
Signed-off-by: Masahisa Kojima <masahisa.kojima@linaro.org>