llvm: Add support for ARMv8-A
This change allows to build the SCP using LLVM for ARMv8-A
targets. The build system now calls the lld linker directly
for any LLVM build. This is due to Clang's preference to use
ld for some ARM targets, e.g. baremetal ARMv8-A. A slight
change was made to the ARMv8-A arch files to instruct lld.

Change-Id: I43f14b09b9b94a8cbfbc9f3f704fa4849d81e0f2
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>