clock: add internal tree configuration
This patch enables tree configuration for clock
management framework. The default configuration is a single
node clock so the tree is disabled for existing platforms.
The newly added tree is not yet used. This will allow to
use a new clock management framework which will be capable
of handling clock nodes dependencies.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I9146cf44169e0ced97379d15b5b5a5f15406a836