clock: Add clock management for setting state
This patch adds set_state function for clock module using
the new clock management framework.
Every clock state change is handled as an asynchronous request
and the response is returned through a delayed event if only
the clock tree is defined. If not it calls driver API
straight away.
At this stage, driver API for set state is only supported
in the synchronous mode.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I04c6cb2fb810b59774bd0e9c01c11b8536eab69c