juno/pvt: Unsubscribe to cluster pd notif when DBGSYS is on
Some sensors are within a cluster power domain which needs to be ON
before the reading can commence. The status of those power domains are
known via the notifications mechanism.

Although the above mechanism works fine in most cases, it is not enough
for Juno. On Juno, the PPU denies the clusters power domain to be
turned OFF if DBGSYS is ON.

This situation is not "reflected" back to subscribers, which would be
left with an improper power status.

This patch allows a simplified mechanism for Juno to work around this
by getting the status of DBGSYS power domain and disable the
notification logic once such domain is ON.

This will allow the above sensors to be read correctly.

Change-Id: I7adaa69ee762db83fd2677450ba8c518811d9677
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>