scmi_clock: Fix clock_id getter
The get_clock_id() helper is there to provide the identifier for the
clock device for each SCMI message. All the messages have an associated
payload and its first field is clock_id.
The only exception to this is CLOCK_SET_RATE message.
The current implementation fo the get_clock_id does not take this
exception in consideration.

Fix get_clock_id() to take into account that not all the payloads have
clock_id as first field.

Change-Id: Ib96a196b1e9b82d7d9fa88e8462530fd844e0af0
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Reported-by: Pierre Gondois <Pierre.Gondois@arm.com>