cmn700: fix num of SYS_CACHE_GRP_HN_NODEID register calc
The number of HN-F CAL instances determines the number of
`SYS_CACHE_GRP_HN_NODEID` registers, not the number of HN-F instances
itself. For calculating the number of `SYS_CACHE_GRP_HN_NODEID`
registers (group_count), instead of using the number of HN-F CAL
instances, number of HN-F instances was used. This patch fixes it to use
number of HN-F CAL instances.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: If3206e86894e17cb6f452ece51653e7d548d8f75