product/rdn2: fix incorrect alignment in config_mhu2
clang-format considers the code to be C++ by default, where the
behaviour of = {} differs from C. clang-format can be forced to parse
the files as C code by ensuring every entry in every initializer ends
with a comma. Add comma after .data = &(...), to force clang-format to
parse the file in C way resulting in desired indentation in MHU2
configuration data.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I719c036b69d78f8e13742c8efecc7d67e5c07853