product/rdn2: fix incorrect alignment in config_smt
clang-format considers the code to be C++ by default, where the
behaviour of = {} differs from C. clang-format can be forced to parse
the files as C code by ensuring every entry in every initializer ends
with a comma. Add comma after .data = &(...), to force clang-format to
parse the file in C way resulting in desired indentation in SMT
configuration data.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I197b933ea0fca6f4b00b71504610ba15c98a8e3e