product/sgm775: Add disabled PSCI protocols w/o resource permissions
Currently for this platform resource permission is enabled and its
resources are managed by the resource permissions module.
At the same time there is no static configuration for PSCI agent should
the resource permission be turned off.
This patch adds a list of protocols that are disabled by default for
PSCI agent when there is no resource permission enabled.

Change-Id: I7544828b2e20b71b090f118b9f343ecf66093735
Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>