product/sgm776: Add disabled PSCI protocols w/o resource permissions
Currently for this platform resource permission is enabled and its
resources are managed by the resource permissions module.
At the same time there is no static configuration for PSCI agent should
the resource permission be turned off.
This patch adds a list of protocols that are disabled by default for
PSCI agent when there is no resource permission enabled.

Change-Id: I6f0477b2421dba1c8f152676abf3ae58e4c55ef4
Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>