cmake: Add Clang-Format integration
This commit introduces the following targets:

- `format-diff-c`: automatically formats modified C/C++ source files
- `format-c`: automatically formats all C/C++ source files
- `check-c`: checks the formatting of all C/C++ source files

These are executed implicitly as part of the `format-diff`, `format` and
`check` targets.

Clang-Format can be found at:
https://clang.llvm.org/docs/ClangFormat.html

Also included in this commit is a submodule for `run-clang-format`, a
Python-based wrapper script for executing CI-like commands with
Clang-Format, which we use to implement the `check-c` target.

run-clang-format can be found at:
https://github.com/Sarcasm/run-clang-format

Change-Id: I72ec079606e2d0f14a54f1a65f44524cbd2a5de4
Signed-off-by: Chris Kay <chris.kay@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>