ppu_v1:Assigning opmode from product configs
This commit allows products to accomodate dynamic
assignment of opmode values. This change was made
to expose operating modes of the ppu such that
they can be configured from each platform. Thus,
every product should have this attribute assigned.
Default opmode is 0 for all platforms and memory
allocated using calloc is zeroed, so all of the
ppu config files are unchanged.

Signed-off-by: Arnold <arnoldgabriel.benedict@arm.com>
Change-Id: Idc7abb7bcb32e1f1125f60d46153be49c69691d7