arch: Remove infinite loop from arch_init
Some architectures don't want to loop indefinitly in the same context of
the initialization. Let architecture decides if it wants to stay in the
initialization context or finish processing the events raised during
init and return.

No functional changes expected for current arch.

fwk_process_event_queue() can then be called to process the pending events.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>