product/rdv1: add config data for mcp romfw bootloader module
The bootloader module is used to load the MCP RAM firmware from the NOR
flash to MCP RAM. Provide the base address and size of the firmware
image to be loaded from and the base address of the location to which
the firmware is loaded to. In addition to this, some of the macros are
renamed to correspondingly match with the macro names used in the SCP
firmware of rdv1 platform.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I23d18ed75f29fffc756bef247326dc0fbf10a454