thermal_mgmt: Add a new module for thermal management
The temperature of the components within a SoC is proportional to the
power that is delivered to (or consumed by) them.
Most of the time, a SoC needs to be maintained within a thermal range.
Affecting the power consumed by some of its devices will help
controlling the temperature.

A thermal management module should be able to affect the final
performance for the platform, thus its place should be in between the
the SCMI performance agent and the performance level driver (DVFS).

The f/w already supports a flexible mechanism to add modules in that
context (performance plugins).

This patch adds a basic Thermal Management module which will operate
as a temperature controller and power regulator for the actors
(CPU, GPU, etc) concerned for the platform.

This module will periodically sample the temperature and update
an internal Proportional-Integral (PI) loop which will feed a
power divider to allocate a power budget for each interested
actor. The coefficients for the PI loop are configurable and all
the tunable parameters are explained in the doc section of this
module.

Change-Id: I2fc0923c2652b4a1d96b26c20f5f85ff8fe01166
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>