cmake: Add CPPCheck support
This commit adds support for CPPCheck, the static analyzer for C and C++
source files. It's run as part of the build process, and so analysis
happens in parallel with building. This gives the static analyzer the
widest view possible of the source files it parses, as it receives a
consistent set of flags and preprocessor, rather than attempting to try
every combination.

Change-Id: If53744f33f1a20e04dd0098827c3dbc5e88ee47f
Signed-off-by: Chris Kay <chris.kay@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>