dvfs: Do not start a new request if already in ctx
When a new set_level is about to start, either a new request or a pending
request is transferred into the current request for processing. But there
are no checks in place to find out whether the candidate request has
already been processed. This can lead to the same request processed twice.

Add a check for discarding duplicate requests in dvfs_set_level_start().

To avoid duplication of code, add also a new helper for request cleanup.

Change-Id: I9e2dfeb4bc3413fd6f66c6eb1c91941601efb7c5
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>