product/rdn2: add support for variant 1 of rdn2 platform
Add support for RD-N2-Cfg1 platform. It is a variant of RD-N2 platform
with a reduced 3x3 mesh and core count reduced to 8. For this variant,
the RD-N2 platform support is reused and for reducing the platform
resources as per the RD-N2-Cfg1 platform configuration, the compile
time macro PLATFORM_VARIANT is used. For the RD-N2-Cfg1 platform, the
PLATFORM_VARIANT value should be set to 1 at compile time.

Change-Id: Ic8e3c2d3068fc2c7e0a9e4601a6a25f73705d9d1
Signed-off-by: Aditya Angadi <aditya.angadi@arm.com>
Signed-off-by: Zakaria Zahi <zakaria.zahi@arm.com>