dvfs: Unconditionally send limits/level updates to scmi_perf
DVFS module provides a configurable way to share performance updates to a
chosen entity. All the platforms are configured such that all the updates are
provided to SCMI Performance.

Given that SCMI Performance is the only interested entity for dvfs updates,
and upcoming changes make this a requirement, remove the optional choice of
recipient and unconditionally send updates to SCMI Performance.

As a consequence, the perf_updated_api is removed as none of the platforms
is configured with updates for other modules than SCMI Performance.

Change-Id: Iaba4f2b3f54edec5326037a71f88289638902906
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>