scmi_perf/dvfs: Move dvfs set/get limits logics into scmi_perf
The performance limits are currently handled in the dvfs module. While this
has the advantage to perform some checks in-place when a new performance level
is requested, it does have some disadvantages. The biggest is that the logic
for limits is missing at SCMI Perf level, where it could be possible to make
early decisions even before having to request a new level to DVFS.

With the upcoming changes in the SCMI Perf layer to make it an interface for
all the performance-related handling, it is best to move the limits into it.

This patch moves the performance limits logic from DVFS to SCMI Perf module.

Change-Id: I1aa973018598e75737265e6cbb53ef3b21ca7eae
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>