scmi: Adapt return values for BASE_DISCOVER_AGENT for v2.0
The current return values for base protocol BASE_DISCOVER_AGENT is
compliant with SCMI v1.0. The version v2.0 of the spec has introduced
an additional agent_id value (return value). Also, the spec v2.0 allows
for the identifier of the agent whose identification is requested to be
equal to 0xFFFFFFFF as a special case.

In preparation to support SCMI v2.0, this patch provides the changes
in both parameters and return values to comply with it.

Change-Id: Ie7e13440890fd33975968c8c78756236f4c83b47
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>