fwk: Add stop sequence
When the scp is running as an application or sub-system, there is a need to
terminate the process gracefully. Let introduce the deinit and stop
sequence to allow the modules and their elements to cleanup and release
resources before terminating the process

Add a new fwk_arch_deinit() to loop on modules and stop them.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>