product/tc1: add configuration data for power domain and PPU modules
Provide the config data such as PPU base addresses and irq map for the
PPU driver and config data for power domain HAL to construct the power
domain tree. Only a single core (core 0) needs to be provided for PPU
config in ROM as AP firmware only uses primary core but all the cores
need to be provided in PPU RAM firmware which is used when the kernel
switches ON secondary cores.

Signed-off-by: Usama Arif <usama.arif@arm.com>
Change-Id: I35386f200e8aad760dae462ef95c4cd77bfa56f4