cmn700: check alignment base and size only for legacy cmn mode
RNSAM region base address should be aligned to size of the region only
in the legacy cmn mode (configurable base address & region size). In
case of configurable start and end address, there is no restriction on
the size of the region. So check for base and size alignment only in the
legacy mode.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Id0fbbf7f74941285405798fd420a022130f91428