product/rdn2: add config data for pcie integ control module
Add configuration data for the PCIe control module to configure the
registers of the first four instances of the IO Macro. The x16 root port
instance in the first four instances of IO macro (first two instances in
case of platform variant 1) are enabled and the address used are matched
with the ECAM and MMIO addresses used in the CMN-700 configuration data.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Signed-off-by: Zakaria Zahi <zakaria.zahi@arm.com>
Change-Id: I5113abe28e261605b8f09888c21d102ce95481f0