cmn700: add hierarchical hashing support with 3 SN mode
CMN-700 support two levels of hashing from RNSAM to HNSAM known as
hierarchical hashing. When this mode is enabled, HN-Fs in the mesh are
grouped evenly forming clusters of HN-Fs. Addresses from the RNSAM will
be hashed to select a cluster of HN-F (first level) and hashed again to
select a HN-F within the cluster (second level). This mode has to be
used along with hashed region (3/5/6-SN mode) in the HNSAM.

Add support to enable hierarchical hashing mode in RNSAM table along
with 3-SN mode in HNSAM table. Other hashed region modes in HNSAM (5-SN
and 6-SN) are not supported in this patch.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I57087b73cb8c75cedcde8249da4f470a3ca44037