product/rdn2: add all NCI GPV regions in cmn700 config
RD-N2 has a total of five instance of NCI. Add NCI Global Programmers
View (GPV) regions for the remaining four NCI instances which are
connected to HN-P nodes in the CMN-700 configuration data. For the
RD-N2-Cfg1 variant, limit the number of NCI GPV regions to two as there
are only two instances of the IO macro on the platform.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Iadff665cb4d940533fba5122e487e62ccb2f54cc