armv7-m: Fix IRQn type conversion issues
On platforms with more than 128 interrupts, IRQ numbers 128 and above
cause negative indexing into the newly-created vector table, due in part
to a newly-introduced `char` cast as well as truncation to `IRQn_Type`
caused by our usage of `-fshort-enums`, causing difficult-to-diagnose
crashes later in the program.

This commit changes IRQs so that they are represented using `uint32_t`,
which is guaranteed to hold all Armv7-M IRQs. We cannot use `IRQn_Type`
to represent arbitrary IRQ integers, as they may be truncated if short
enums are enabled and the IRQ has not been included in the type.

This bug was originally introduced in commit `77e9fe7f`.

Change-Id: I398cc922521d78a2498862b75d85162e7f034d09
Signed-off-by: Chris Kay <chris.kay@arm.com>