fwk_log: Conditionally remove arguments from log macros
Currently when the chosen LOG_LEVEL is higher than the level for the
macro FWK_LOG_<LEVEL> the arguments are being voided and the log is not
printed.

Those strings remain in the built image though, taking unnecessary space.

This patch entirely removes the arguments from the macro and as a
consequence the strings are removed saving memory.

Change-Id: I32d2ca8e42ac3cb8e7bcc87e49431a857e07c452
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>