perf_plugins_handler: Add support for full data snapshot
The data received by the plugins on a periodic tick is a set of performance
level and limits coming from the SCMI layer, for a particular performance
domain.
This means that if a plugin wishes to have the data for all the domains
it has to wait for each of the updates to come through. In this scenario,
once the plugin has all the data for all the domains, it is though too
late to affect the performance, since it must be made on each individual
tick/domain.

This limitation can be overcome by providing the whole/full snapshot of
data for all the domains in one go. The plugin handler will then take care
of the aggregation of values.

This patch adds support for achieving the feature above.

Change-Id: Ia758512865d3bf112a551702dd371f970c96e7f3
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>