mod_resource_perms: Fix recursion on mod_res_message_id_to_index
The SCMI spec described that protocol identifiers within the range
0x80-0xFF are reserved for vendor or platform-specific extension.

In the function mod_res_message_id_to_index() if the protocol identifier
falls within the above range, a call to the same function would happen,
causing recursion.

This patch removes the recursion by calling the platform-specific
message-to-index function instead.

Change-Id: Idadac88443c2eccf0fa138495e16e19f3d739c7e
Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>