framework: Move async_response_event variable initialization.
For non multi thread builds, the function __fwk_thread_run
calls process_next_event for every event, or even for no event.
The process_next_event uses async_response_event and initializes
everytime it gets called. However this initialization is not
required for every call. This change moves this initialization
near to the local scope where it is necessary, saving cycles
for the  calls that do not require async_response_event
initialization.

Change-Id: If6543034df4ca4db43bc0a1ea944f2ac6a5d91e8
Signed-off-by: Girish Pathak <girish.pathak@arm.com>