framework: Add light events.
At present 'struct fwk_event' is large because of the field params.
Since 'struct fwk_event' is often instantiated on the stack and
initialized, the CPU cycles are wasted for the fields which are
not needed for events that don't need 'params'. This change addresses
this issue with the introduction of 'struct fwk_event_light'.

This new structure is essentially similar to 'struct fwk_event'
but without 'params' field and few others. Hence fewer CPU cycles
are required to initialize these objects, improving performance of the
runtime use cases such as DVFS.

Note, the light events are not intended to change the behavior of the
framework's event functionality, instead it provides an additional way
of sending smaller events from one module to another by reducing CPU
cycle overhead at the expense of missing 'params' field.
In the framework, light events are converted to usual events and are
processed by module's 'process_event' function.

Change-Id: Iba79a918ada95dddf9850f601c518482525a08cc
Signed-off-by: Girish Pathak <girish.pathak@arm.com>