mod_scmi: Use light events.
In the function signal_message, object of struct fwk_event
is fully instantiated, which at present is not required.
This change replaces this object with struct fwk_event_light
object, which is small and require fewer cycles to instantiate.
Since this function is called very often at the runtime,
saving of CPU cycles in this function increases the overall
runtime performance of the SCMI use cases.

Change-Id: Id36602055b622642230374bef913f92254b65b33
Signed-off-by: Girish Pathak <girish.pathak@arm.com>