cmn700: drop get_root_node helper function
The get_root_node function helps to get the Root Node base address based
on the position (node id) of the HND node and the mesh size. CMN-700
aliases the root node base at the PERIPHBASE address (base address of
the CMN-700) so any access to the PERIPHBASE address will be aliased to
Root Node base. Hence the calculation implemented with the get_root_node
function is no longer required.

Create a separate function for calculating encoding and masking bits
which were done as part of get_root_node function.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I876d0c7eb1974f109be3345cc95600b66bd499b1